pyproject.toml
[tool.poetry]
name = "ksabee-backend"
version = "0.1.0"
description = ""
authors = ["SnowSuno <snowsuno.dev@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.73.0"
uvicorn = "^0.17.4"
pydantic = { extras = ["dotenv"], version = "^1.9.0" }
tortoise-orm = "^0.18.1"
aiohttp = "^3.8.1"
bs4 = "^0.0.1"
asyncpg = "^0.25.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
